+++ ving/utils/int2str.mjs
@@ -1,15 +1,15 @@
 import { encode, decode } from './base62.mjs';
 import { encrypt, decrypt } from './skipjack.mjs';
-import { isNumber } from '#ving/utils/identify.mjs';
+import { isNumber, isString } from '#ving/utils/identify.mjs';
 
 /**
  * Generates an encrypted string from an integer. See the environment variable `VING_SKIPJACK_KEY`.
- * @param {number} int The number to be encrypted.
+ * @param {number} int The number to be encrypted. If you pass a string it will not be encoded, but just return the string passed without encrypting becasue we're assuming it is already an encrypted string.
  * @returns {string} The resulting encrypted string. Note that this will be different for different sites based upon what is in `VING_SKIPJACK_KEY`.
  * @example
  * stringifyId(1) // vJH34sj3
  */
-export const stringifyId = (int) => encode(encrypt(int));
+export const stringifyId = (int) => isString(int) ? int : encode(encrypt(int));
 
 /**
  * Unencrypts a string back to its original integer value.